Clean jtop functions
